+++ src/BlockHorizons/BlockSniper/brush/types/RegenerateType.php
@@ -5,6 +5,9 @@ declare(strict_types=1);
 namespace BlockHorizons\BlockSniper\brush\types;
 
 use BlockHorizons\BlockSniper\brush\BaseType;
+use BlockHorizons\BlockSniper\revert\async\AsyncUndo;
+use BlockHorizons\BlockSniper\sessions\Session;
+use BlockHorizons\BlockSniper\sessions\SessionManager;
 use pocketmine\level\ChunkManager;
 use pocketmine\level\format\Chunk;
 use pocketmine\math\Vector3;
@@ -12,16 +15,19 @@ use pocketmine\Player;
 
 /*
  * Regenerates the chunk looked at.
- * This brush can NOT undo.
  */
 
 class RegenerateType extends BaseType{
 
 	public const ID = self::TYPE_REGENERATE;
 
+	/** @var Session */
+	private $session;
+
 	public function __construct(Player $player, ChunkManager $manager, \Generator $blocks = null){
 		parent::__construct($player, $manager, $blocks);
 		$this->center = $player->getTargetBlock(100)->asVector3();
+		$this->session = SessionManager::getPlayerSession($player);
 	}
 
 	/**
@@ -34,9 +40,13 @@ class RegenerateType extends BaseType{
 		$x = $this->center->x >> 4;
 		$z = $this->center->z >> 4;
 
-		$this->getLevel()->setChunk($x, $z, new Chunk($x, $z));
+		$oldChunk = $this->getLevel()->getChunk($x, $z);
+		$c = new Chunk($x, $z);
+		$this->getLevel()->setChunk($x, $z, $c);
 
 		$this->getLevel()->populateChunk($x, $z, true);
+
+		$this->session->getRevertStore()->saveRevert(new AsyncUndo([$c], [$oldChunk], $this->session->getSessionOwner()->getName(), $this->level));
 
 		if(false){
 			// Make PHP recognize this is a generator.